Sample markdown page of a 1st chapter.
This is a temporary sample page to hold info about the first chapter of Essentials of Linear Algebra by 3Blue1Brown. In the next few hours, it will be transferred into an html+css single-page website that will show the representation of my capability (or incapability) of doing the responsibilities of the internship.